bigtable/demo/requirements.txt

google-cloud-bigtable==2.23.1